Updated models to use CMS for programs and courses
Updated the model definitions and database structure to allow for using
the CMS to manage creation and contents of Program and Course
objects. This allows for stakeholders in the respective departments to
handle the messaging on program pages in a self-service manner. Access
rights to the different program pages can be managed with the Wagtail
authentication and authorization mechanisms.